[JIT][Symbolic Shape Analysis] expose op shape functions
[ghstack-poisoned]
