+++ torch/__init__.py
@@ -833,6 +833,7 @@ def compiled_with_cxx11_abi():
 # Import the ops "namespace"
 from torch._ops import ops
 from torch._classes import classes
+from torch import shape as shape
 
 # quantization depends on torch.fx
 # Import quantization
