+++ torch/shape/__init__.py
@@ -1,0 +1,1 @@
+from . import _ops as ops
